ref: 拆分 联网通知 & 网络鉴权 & 更新
