+++ Plain Craft Launcher 2/Modules/Updates/UpdateEnums.cs
@@ -1,0 +1,19 @@
+﻿namespace PCL;
+
+public class UpdateEnums
+{
+    public enum VersionStatus
+    {
+        Latest,
+        NotLatest,
+        Unknown
+    }
+    
+    public enum UpdateType
+    {
+        Silent = 0,
+        PromptOnly = 1,
+        DownloadAndPrompt = 2,
+        UpdateNow = 3
+    }
+}
