+++ Plain Craft Launcher 2/Pages/PageSetup/PageSetupUpdate.xaml.cs
@@ -35,8 +35,8 @@ public partial class PageSetupUpdate
         {
             // 修复：使用 dynamic 绕过命名空间重名导致的编译期类型冲突，
             // 或者你可以尝试替换为 PCL.Core.App.SemVer.Parse(ModBase.VersionBaseName)
-            if (await ModSecret.RemoteServer.IsLatestAsync(
-                    ModSecret.IsCurrentVersionBeta ? UpdateChannel.beta : UpdateChannel.stable,
+            if (await UpdateManager.RemoteServer.IsLatestAsync(
+                    UpdateManager.IsCurrentVersionBeta ? UpdateChannel.beta : UpdateChannel.stable,
                     ModBase.IsArm64System ? UpdateArch.arm64 : UpdateArch.x64,
                     SemVer.Parse(ModBase.VersionBaseName),
                     ModBase.VersionCode))
@@ -69,8 +69,8 @@ public partial class PageSetupUpdate
                 Exception checkUpdateEx = null;
                 try
                 {
-                    UpdateInfo = ModSecret.RemoteServer.GetLatestVersion(
-                        ModSecret.IsCurrentVersionBeta
+                    UpdateInfo = UpdateManager.RemoteServer.GetLatestVersion(
+                        UpdateManager.IsCurrentVersionBeta
                             ? UpdateChannel.beta
                             : UpdateChannel.stable, ModBase.IsArm64System ? UpdateArch.arm64 : UpdateArch.x64);
                     TextUpdateName.Text = "PCL CE " + VersionNameFormat(UpdateInfo.VersionName);
@@ -96,12 +96,12 @@ public partial class PageSetupUpdate
                     return;
                 }
 
-                if (ModSecret.UpdateLoader is not null && ModSecret.UpdateLoader.State == ModBase.LoadState.Loading)
+                if (UpdateManager.UpdateLoader is not null && UpdateManager.UpdateLoader.State == ModBase.LoadState.Loading)
                 {
                     BtnUpdate_Timer();
                     BtnUpdate.IsEnabled = false;
                 }
-                else if (ModSecret.IsUpdateWaitingRestart)
+                else if (UpdateManager.IsUpdateWaitingRestart)
                 {
                     BtnUpdate.Text = "重启安装";
                     BtnUpdate.IsEnabled = true;
@@ -137,9 +137,9 @@ public partial class PageSetupUpdate
 
     public void BtnUpdate_Timer()
     {
-        while (ModSecret.UpdateLoader is not null && ModSecret.UpdateLoader.State == ModBase.LoadState.Loading)
+        while (UpdateManager.UpdateLoader is not null && UpdateManager.UpdateLoader.State == ModBase.LoadState.Loading)
         {
-            ModBase.RunInUi(() => BtnUpdate.Text = $"{Math.Round(ModSecret.UpdateLoader.Progress, 2)}%");
+            ModBase.RunInUi(() => BtnUpdate.Text = $"{Math.Round(UpdateManager.UpdateLoader.Progress, 2)}%");
             Thread.Sleep(200);
         }
     }
@@ -158,9 +158,9 @@ public partial class PageSetupUpdate
             return;
         }
 
-        if (ModSecret.IsUpdateWaitingRestart) ModSecret.UpdateRestart(true);
+        if (UpdateManager.IsUpdateWaitingRestart) UpdateManager.UpdateRestart(true);
         // 开始更新流程
-        ModSecret.UpdateStart(ModSecret.UpdateType.UpdateNow);
+        UpdateManager.UpdateStart(UpdateEnums.UpdateType.UpdateNow);
     }
 
     private void BtnChangelogDetail_Click(object sender, EventArgs e)
